Refactor auditor.py to reduce ACL by moving helpers to auditor_utils.py
- Extracted `_collect_directory_stats`, `_get_critical_files_content`, `_get_project_signatures`, `_check_agents_md`, `_check_linter_config`, `_check_lock_file`, and `_validate_pyproject` to `src/agent_scorecard/auditor_utils.py`.
- Updated `src/agent_scorecard/auditor.py` to import these helper functions.
- Verified ACL score improvement for `auditor.py` (20 -> 80).
- Passed all tests and linting.
